Take into account that input strings can contain surrogate pairs
If Liblouis is compiled for UCS-2, this doesn't change anything,
except that characterAttributes, interCharacterAttributes and typeform
are not supported when the input has surrogate pairs.

If Liblouis is compiled for UCS-4, this commit fixes an encoding
error when surrogate pairs are present.

Also fixed bugs in the decoding of braille using
StandardDisplayTable.UNICODE (with UCS-4 Liblouis) or custom
DisplayTables.